whole new world (#116)
New static analysis (courtesy of https://github.com/toshok/echojs-maam), compiler IR, optimization passes, and GC implementation.  documentation to come.

+++ runtime/ejs-types.h
@@ -27,7 +27,29 @@ typedef double jsdouble;
 
 typedef uint16_t jschar;
 
-typedef uint32_t GCObjectHeader;
+// The object header, widened to 64 bits as one joint GC/shapes layout
+// (written once — see docs/gc-plan.md "Object header, forwarding, and
+// shapes" and docs/shapes-plan.md "Object layout, in two steps"):
+//
+//   bits  0-31  the pre-existing 32-bit header: EJSScanType in the low
+//               bits, user flags at EJS_GC_USER_FLAGS_SHIFT (unchanged)
+//   bits 32-55  shape index (0 = dictionary mode / untracked)
+//   bit  56     shaped-storage mode bit
+//   bit  57     YOUNG — allocated since the last collection (profiling
+//               profiling; a nursery age bit in waiting)
+//   bit  58     PINNED — conservatively referenced this cycle (profiling
+//               profiling, cleared each cycle)
+//   bit  59     FORWARDED — the word is a forwarding record, not a
+//               header: target address in bits 0-46 (see
+//               ejs-gc.h _ejs_gc_forward)
+//   bit  60     DIRTY — the object is in the generational remembered
+//               buffer (object-remembering write barrier)
+//   bits 61-63  reserved for the GC (future mark/card bits)
+//
+// EJSObject absorbs the widening into what was padding (sizeof
+// unchanged); EJSPrimString/EJSPrimSymbol keep their sizes; EJSClosureEnv
+// grows by 8.  lib/types.ts mirrors this in the same commit.
+typedef uint64_t GCObjectHeader;
 
 #if defined(__GNUC__) && (__GNUC__ > 2)
 # define EJS_LIKELY(x)   (__builtin_expect((x), 1))
